Reformat complete
Code reformatted.
Motion functionblocks are now called in one lump together at the start of the drive state machine.
All variables have been renamed in the AxisControl structure to follow a better format.
The done latch works consistently now as it is no longer needing to use a rising edge trigger (as the done will come on for 1 cycle even if execute not held high which I was not aware of).
Visualisation has been updated to match new variables and format, gearing and multi-gearing have not been tested since update but everything else implemented seems to be working correctly.
There are some quirks that need to be looked in to specifically that I was originally using the "HasJob" flag of the axes to assess when it was okay to issue new commands however writing parameters does not count as a "hasjob" flag which affected the logic. Alternative has been put in but needs to be reviewed (see lines 116 to 121 of AxDriveFB).